refactor(components): split HealthDashboardPage into per-section subcomponents
Part of #290
