+++ src/crosshook-native/src/components/pages/HealthDashboardPage.tsx
@@ -5,821 +5,27 @@ import { useOfflineReadiness } from '../../hooks/useOfflineReadiness';
 import type { TrendDirection } from '../../hooks/useProfileHealth';
 import { useProtonMigration } from '../../hooks/useProtonMigration';
 import { useVersionCheck } from '../../hooks/useVersionCheck';
-import type { MigrationSuggestion, OfflineReadinessReport, ProtonPathField } from '../../types';
-import type { EnrichedProfileHealthReport, HealthIssue, HealthStatus } from '../../types/health';
-import type { VersionCorrelationStatus } from '../../types/version';
 import { formatRelativeTime } from '../../utils/format';
 import { HealthBadge } from '../HealthBadge';
 import { RouteBanner } from '../layout/RouteBanner';
 import type { AppRoute } from '../layout/Sidebar';
 import { MigrationReviewModal } from '../MigrationReviewModal';
-import { OfflineReadinessPanel } from '../OfflineReadinessPanel';
 import { OfflineStatusBadge } from '../OfflineStatusBadge';
-import { CollapsibleSection } from '../ui/CollapsibleSection';
-
-type IssueCategory =
-  | 'missing_executable'
-  | 'missing_trainer'
-  | 'missing_dll'
-  | 'missing_proton'
-  | 'missing_compatdata'
-  | 'missing_prefix'
-  | 'inaccessible_path'
-  | 'other';
-
-interface IssueCategoryCount {
-  category: IssueCategory;
-  label: string;
-  count: number;
-}
-
-const CATEGORY_LABELS: Record<IssueCategory, string> = {
-  missing_executable: 'Missing executable',
-  missing_trainer: 'Missing trainer',
-  missing_dll: 'Missing DLL',
-  missing_proton: 'Missing/invalid Proton path',
-  missing_compatdata: 'Inaccessible compatdata',
-  missing_prefix: 'Missing prefix path',
-  inaccessible_path: 'Inaccessible path',
-  other: 'Other',
-};
-
-function categorizeIssue(issue: HealthIssue): IssueCategory {
-  const { field, severity } = issue;
-  if (field === 'game.executable_path') return 'missing_executable';
-  if (field === 'trainer.path') return 'missing_trainer';
-  if (field.startsWith('injection.dll_paths')) return 'missing_dll';
-  if (field === 'steam.proton_path' || field === 'runtime.proton_path') return 'missing_proton';
-  if (field === 'steam.compatdata_path') return 'missing_compatdata';
-  if (field === 'runtime.prefix_path') return 'missing_prefix';
-  if (severity === 'warning') return 'inaccessible_path';
-  return 'other';
-}
-
-type SortField =
-  | 'name'
-  | 'status'
-  | 'issues'
-  | 'last_success'
-  | 'launch_method'
-  | 'failures'
-  | 'favorite'
-  | 'version_status'
-  | 'offline_score';
-type SortDirection = 'asc' | 'desc';
-type StatusFilter = 'all' | HealthStatus;
-
-const STATUS_RANK: Record<string, number> = { broken: 2, stale: 1, healthy: 0 };
-
-const VERSION_STATUS_RANK: Partial<Record<VersionCorrelationStatus, number>> = {
-  both_changed: 3,
-  game_updated: 2,
-  trainer_changed: 2,
-  matched: 1,
-  update_in_progress: 0,
-  untracked: -1,
-  unknown: -1,
-};
-
-function getVersionStatusColor(status: VersionCorrelationStatus | null | undefined): string {
-  if (status === 'matched') return 'var(--crosshook-color-success)';
-  if (status === 'game_updated' || status === 'trainer_changed' || status === 'both_changed') {
-    return 'var(--crosshook-color-warning)';
-  }
-  return 'var(--crosshook-color-text-subtle)';
-}
-
-function getVersionStatusLabel(status: VersionCorrelationStatus | null | undefined): string {
-  switch (status) {
-    case 'matched':
-      return 'Matched';
-    case 'game_updated':
-      return 'Game Updated';
-    case 'trainer_changed':
-      return 'Trainer Changed';
-    case 'both_changed':
-      return 'Both Changed';
-    case 'update_in_progress':
-      return 'Updating';
-    case 'untracked':
-      return 'Untracked';
-    case 'unknown':
-      return 'Unknown';
-    default:
-      return 'Untracked';
-  }
-}
-
-type CardTrend = 'up' | 'down' | null;
-
-function mergeOfflineReadinessForRow(
-  report: EnrichedProfileHealthReport,
-  hookReport: OfflineReadinessReport | undefined
-): OfflineReadinessReport | undefined {
-  const brief = report.offline_readiness;
-  const offlineIssues = report.issues.filter((i) => i.field.startsWith('offline_readiness.'));
-  const checks = offlineIssues.map((i) => ({
-    ...i,
-    field: i.field.replace(/^offline_readiness\./, ''),
-  }));
-  if (brief) {
-    return {
-      profile_name: brief.profile_name,
-      score: brief.score,
-      readiness_state: brief.readiness_state,
-      trainer_type: brief.trainer_type,
-      blocking_reasons: [...brief.blocking_reasons],
-      checked_at: brief.checked_at,
-      checks,
-    };
-  }
-  if (hookReport) {
-    return hookReport;
-  }
-  if (checks.length > 0) {
-    return {
-      profile_name: report.name,
-      score: 0,
-      readiness_state: 'unknown',
-      trainer_type: 'unknown',
-      blocking_reasons: [],
-      checked_at: report.checked_at,
-      checks,
-    };
-  }
-  return undefined;
-}
-
-function offlineSortScore(report: EnrichedProfileHealthReport, hookReport: OfflineReadinessReport | undefined): number {
-  const merged = mergeOfflineReadinessForRow(report, hookReport);
-  if (merged && merged.score !== undefined && !Number.isNaN(merged.score)) {
-    return merged.score;
-  }
-  return -1;
-}
-
-function TrendArrow({ trend, improving }: { trend: CardTrend; improving: boolean }) {
-  if (trend === null) return null;
-  const isPositive = (trend === 'up' && improving) || (trend === 'down' && !improving);
-  const color = isPositive ? 'var(--crosshook-color-success)' : 'var(--crosshook-color-danger)';
-  const label = trend === 'up' ? 'trending up' : 'trending down';
-  return (
-    <span
-      role="img"
-      aria-label={label}
-      style={{ fontSize: '0.8em', fontWeight: 700, color, marginLeft: '4px', lineHeight: 1 }}
-    >
-      {trend === 'up' ? '\u2191' : '\u2193'}
-    </span>
-  );
-}
-
-function SummaryCard({
-  count,
-  label,
-  accentColor,
-  disabled,
-  trend,
-  improving,
-}: {
-  count: number | null;
-  label: string;
-  accentColor: string;
-  disabled?: boolean;
-  trend?: CardTrend;
-  improving?: boolean;
-}) {
-  const displayCount = disabled || count === null ? '—' : String(count);
-  return (
-    <div
-      className="crosshook-card crosshook-health-dashboard-card"
-      style={{ borderLeftColor: accentColor }}
-      aria-disabled={disabled}
-    >
-      <div className="crosshook-health-dashboard-card__count" style={{ color: disabled ? undefined : accentColor }}>
-        {displayCount}
-        {!disabled && trend != null && <TrendArrow trend={trend} improving={improving ?? false} />}
-      </div>
-      <div className="crosshook-health-dashboard-card__label crosshook-muted">{label}</div>
-    </div>
-  );
-}
-
-function TableToolbar({
-  statusFilter,
-  onStatusFilter,
-  searchQuery,
-  onSearchQuery,
-  shownCount,
-  totalCount,
-  loading,
-  onRecheck,
-  lastValidated,
-  missingProtonCount,
-  onFixProtonPaths,
-  isScanning,
-  onCheckAllVersions,
-  isVersionScanning,
-  versionScanProgress,
-}: {
-  statusFilter: StatusFilter;
-  onStatusFilter: (f: StatusFilter) => void;
-  searchQuery: string;
-  onSearchQuery: (q: string) => void;
-  shownCount: number;
-  totalCount: number;
-  loading: boolean;
-  onRecheck: () => void;
-  lastValidated: string | null;
-  missingProtonCount?: number;
-  onFixProtonPaths?: () => void;
-  isScanning?: boolean;
-  onCheckAllVersions?: () => void;
-  isVersionScanning?: boolean;
-  versionScanProgress?: { done: number; total: number } | null;
-}) {
-  const statusOptions: { value: StatusFilter; label: string }[] = [
-    { value: 'all', label: 'All' },
-    { value: 'healthy', label: 'Healthy' },
-    { value: 'stale', label: 'Stale' },
-    { value: 'broken', label: 'Broken' },
-  ];
-
-  return (
-    <div className="crosshook-health-dashboard-toolbar">
-      <fieldset className="crosshook-health-dashboard-toolbar__filters crosshook-fieldset-reset">
-        <legend className="crosshook-visually-hidden">Filter by status</legend>
-        {statusOptions.map((opt) => (
-          <button
-            key={opt.value}
-            type="button"
-            className={`crosshook-status-chip crosshook-health-dashboard-toolbar__pill${statusFilter === opt.value ? ' crosshook-health-dashboard-toolbar__pill--active' : ''}`}
-            onClick={() => onStatusFilter(opt.value)}
-            aria-pressed={statusFilter === opt.value}
-          >
-            {opt.label}
-          </button>
-        ))}
-      </fieldset>
-      <input
-        type="search"
-        className="crosshook-input crosshook-health-dashboard-toolbar__search"
-        placeholder="Filter profiles..."
-        value={searchQuery}
-        maxLength={200}
-        onChange={(e) => onSearchQuery(e.target.value)}
-        aria-label="Filter profiles by name"
-      />
-      <span className="crosshook-muted crosshook-health-dashboard-toolbar__count">
-        Showing {shownCount} of {totalCount}
-      </span>
-      <div className="crosshook-health-dashboard-toolbar__recheck">
-        {lastValidated && (
-          <span className="crosshook-muted crosshook-health-dashboard-toolbar__validated">
-            {formatRelativeTime(lastValidated)}
-          </span>
-        )}
-        <button
-          type="button"
-          className="crosshook-button crosshook-button--ghost"
-          disabled={loading}
-          onClick={onRecheck}
-          aria-label="Re-check all profiles"
-        >
-          {loading ? '↻ Checking...' : '↻ Re-check All'}
-        </button>
-        {onCheckAllVersions !== undefined && (
-          <button
-            type="button"
-            className="crosshook-button crosshook-button--ghost crosshook-focus-ring crosshook-nav-target"
-            style={{ minHeight: 'var(--crosshook-touch-target-min)' }}
-            disabled={isVersionScanning}
-            onClick={onCheckAllVersions}
-            aria-label="Check version status for all displayed profiles"
-            aria-disabled={isVersionScanning}
-          >
-            {isVersionScanning
-              ? versionScanProgress
-                ? `Checking ${versionScanProgress.done}/${versionScanProgress.total}\u2026`
-                : 'Checking\u2026'
-              : 'Check All Versions'}
-          </button>
-        )}
-        {missingProtonCount !== undefined && missingProtonCount >= 2 && onFixProtonPaths !== undefined && (
-          <button
-            type="button"
-            className="crosshook-button crosshook-focus-ring crosshook-nav-target"
-            style={{ minHeight: 'var(--crosshook-touch-target-min)' }}
-            disabled={isScanning}
-            onClick={onFixProtonPaths}
-            aria-label={`Fix ${missingProtonCount} profiles with stale Proton paths`}
-            aria-disabled={isScanning}
-          >
-            {isScanning ? 'Scanning\u2026' : `Fix Proton Paths (${missingProtonCount})`}
-          </button>
-        )}
-      </div>
-    </div>
-  );
-}
-
-function SortArrow({
-  field,
-  sortField,
-  sortDirection,
-}: {
-  field: SortField;
-  sortField: SortField;
-  sortDirection: SortDirection;
-}) {
-  if (field !== sortField)
-    return (
-      <span
-        className="crosshook-health-dashboard-sort-arrow crosshook-health-dashboard-sort-arrow--inactive"
-        aria-hidden="true"
-      >
-        ↕
-      </span>
-    );
-  return (
-    <span
-      className="crosshook-health-dashboard-sort-arrow crosshook-health-dashboard-sort-arrow--active"
-      aria-hidden="true"
-    >
-      {sortDirection === 'asc' ? '↑' : '↓'}
-    </span>
-  );
-}
-
-function SkeletonCards() {
-  return (
-    <>
-      {[0, 1, 2, 3].map((i) => (
-        <div key={i} className="crosshook-card crosshook-health-dashboard-skeleton-card" aria-hidden="true">
-          <div className="crosshook-health-dashboard-skeleton crosshook-health-dashboard-skeleton-card__count" />
-          <div className="crosshook-health-dashboard-skeleton crosshook-health-dashboard-skeleton-card__label" />
-        </div>
-      ))}
-    </>
-  );
-}
-
-function _SkeletonRows() {
-  const widths = ['60%', '80%', '70%', '55%', '75%', '65%'];
-  return (
-    <>
-      {[0, 1, 2, 3, 4, 5].map((i) => (
-        <tr key={i} className="crosshook-health-dashboard-skeleton-row">
-          <td>
-            <span
-              className="crosshook-health-dashboard-skeleton crosshook-health-dashboard-skeleton-cell"
-              style={{ width: '56px' }}
-            />
-          </td>
-          <td>
-            <span
-              className="crosshook-health-dashboard-skeleton crosshook-health-dashboard-skeleton-cell"
-              style={{ width: widths[i] }}
-            />
-          </td>
-          <td>
-            <span
-              className="crosshook-health-dashboard-skeleton crosshook-health-dashboard-skeleton-cell"
-              style={{ width: '24px' }}
-            />
-          </td>
-          <td>
-            <span
-              className="crosshook-health-dashboard-skeleton crosshook-health-dashboard-skeleton-cell"
-              style={{ width: '36px' }}
-            />
-          </td>
-        </tr>
-      ))}
-    </>
-  );
-}
-
-function RecentFailuresPanel({ profiles }: { profiles: EnrichedProfileHealthReport[] }) {
-  if (profiles.length === 0) {
-    return (
-      <CollapsibleSection title="Recent Failures" defaultOpen={false}>
-        <p className="crosshook-muted">No profiles with recent launch failures.</p>
-      </CollapsibleSection>
-    );
-  }
-  return (
-    <CollapsibleSection title="Recent Failures" defaultOpen={false}>
-      <ul className="crosshook-health-dashboard-failures-list">
-        {profiles.map((report) => (
-          <li key={report.name} className="crosshook-health-dashboard-failures-item">
-            <span className="crosshook-health-dashboard-failures-item__name">{report.name}</span>
-            <span className="crosshook-status-chip crosshook-health-dashboard-failures-item__count">
-              {report.metadata?.failure_count_30d} failure{report.metadata?.failure_count_30d !== 1 ? 's' : ''} (30d)
-            </span>
-            <span className="crosshook-muted crosshook-health-dashboard-failures-item__last-success">
-              {report.metadata?.last_success
-                ? `Last success ${formatRelativeTime(report.metadata.last_success)}`
-                : 'No successful launches recorded'}
-            </span>
-          </li>
-        ))}
-      </ul>
-    </CollapsibleSection>
-  );
-}
-
-const DRIFT_STATE_MESSAGES: Record<string, string> = {
-  missing: 'Exported launcher not found',
-  moved: 'Launcher has moved',
-  stale: 'Launcher may be outdated',
-};
-
-function LauncherDriftPanel({ profiles }: { profiles: EnrichedProfileHealthReport[] }) {
-  const driftProfiles = useMemo(() => {
-    return (profiles ?? []).filter((r) => {
-      const state = r.metadata?.launcher_drift_state;
-      return state != null && state !== 'aligned' && state !== 'unknown';
-    });
-  }, [profiles]);
-
-  return (
-    <CollapsibleSection title="Launcher Drift" defaultOpen={false}>
-      {driftProfiles.length === 0 ? (
-        <p className="crosshook-muted">All exported launchers are current.</p>
-      ) : (
-        <ul className="crosshook-health-dashboard-issues-list">
-          {driftProfiles.map((r) => {
-            const state = r.metadata?.launcher_drift_state ?? '';
-            const message = DRIFT_STATE_MESSAGES[state] ?? state;
-            return (
-              <li key={r.name} className="crosshook-health-dashboard-issue">
-                <span className="crosshook-health-dashboard-issue__field">{r.name}</span>
-                <span className="crosshook-health-dashboard-issue__message crosshook-muted">{message}</span>
-              </li>
-            );
-          })}
-        </ul>
-      )}
-    </CollapsibleSection>
-  );
-}
-
-function IssueBreakdownPanel({ profiles }: { profiles: EnrichedProfileHealthReport[] }) {
-  const { categoryCounts, totalRawIssues } = useMemo(() => {
-    const counts = new Map<IssueCategory, number>();
-    let totalRaw = 0;
-    for (const report of profiles) {
-      totalRaw += report.issues.length;
-      for (const issue of report.issues) {
-        if (issue.severity === 'info') {
-          continue;
-        }
-        const cat = categorizeIssue(issue);
-        counts.set(cat, (counts.get(cat) ?? 0) + 1);
-      }
-    }
-    const result: IssueCategoryCount[] = [];
-    for (const [category, count] of counts.entries()) {
-      result.push({ category, label: CATEGORY_LABELS[category], count });
-    }
-    result.sort((a, b) => b.count - a.count);
-    return {
-      categoryCounts: result,
-      totalRawIssues: totalRaw,
-    };
-  }, [profiles]);
-
-  const maxCount = categoryCounts.length > 0 ? categoryCounts[0].count : 1;
-
-  return (
-    <CollapsibleSection title="Issue Breakdown" defaultOpen>
-      {categoryCounts.length === 0 && totalRawIssues === 0 ? (
-        <p className="crosshook-muted">No issues found across all profiles.</p>
-      ) : categoryCounts.length === 0 && totalRawIssues > 0 ? (
-        <p className="crosshook-muted">
-          No actionable issues in this summary. Informational checks (including offline readiness notes) are excluded
-          here — expand a profile row to see full details.
-        </p>
-      ) : (
-        <ul className="crosshook-health-dashboard-breakdown-list">
-          {categoryCounts.map(({ category, label, count }) => (
-            <li key={category} className="crosshook-health-dashboard-breakdown-row">
-              <span className="crosshook-health-dashboard-breakdown-label">{label}</span>
-              <span className="crosshook-status-chip crosshook-health-dashboard-breakdown-badge">{count}</span>
-              <div className="crosshook-health-dashboard-breakdown-bar-track" aria-hidden="true">
-                <div
-                  className="crosshook-health-dashboard-breakdown-bar"
-                  style={{ width: `${Math.round((count / maxCount) * 100)}%` }}
-                />
-              </div>
-            </li>
-          ))}
-        </ul>
-      )}
-    </CollapsibleSection>
-  );
-}
-
-function IssueDetailRow({
-  report,
-  offlineReadinessReport,
-  onRevalidate,
-  onFixNavigate,
-}: {
-  report: EnrichedProfileHealthReport;
-  offlineReadinessReport: OfflineReadinessReport | undefined;
-  onRevalidate: (name: string) => void;
-  onFixNavigate: (name: string) => void | Promise<void>;
-}) {
-  const {
-    scanResult,
-    isScanning,
-    applyResult,
-    isApplying,
-    error: migrationError,
-    scanMigrations,
-    applySingleMigration,
-  } = useProtonMigration();
-  const [activeMigrationField, setActiveMigrationField] = useState<string | null>(null);
-  const [successNotice, setSuccessNotice] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (!applyResult || applyResult.outcome !== 'applied') return;
-    setSuccessNotice('Proton path updated.');
-    setActiveMigrationField(null);
-    const timer = setTimeout(() => setSuccessNotice(null), 5000);
-    return () => clearTimeout(timer);
-  }, [applyResult]);
-
-  function getProtonPathField(issueField: string): ProtonPathField | null {
-    if (issueField === 'steam.proton_path') return 'steam_proton_path';
-    if (issueField === 'runtime.proton_path') return 'runtime_proton_path';
-    return null;
-  }
-
-  async function handleUpdateProton(issueField: string) {
-    setActiveMigrationField(issueField);
-    setSuccessNotice(null);
-    await scanMigrations();
-  }
-
-  async function handleApply(suggestion: MigrationSuggestion) {
-    await applySingleMigration({
-      profile_name: suggestion.profile_name,
-      field: suggestion.field,
-      new_path: suggestion.new_path,
-    });
-  }
-
-  const meta = report.metadata;
-  return (
-    <tr className="crosshook-health-dashboard-expanded-row">
-      <td colSpan={11}>
-        <div className="crosshook-health-dashboard-expanded-content">
-          {offlineReadinessReport ? (
-            <div className="crosshook-panel" style={{ marginBottom: 12 }}>
-              <h3 className="crosshook-heading-section" style={{ margin: '0 0 8px', fontSize: '1rem' }}>
-                Offline readiness
-              </h3>
-              <OfflineReadinessPanel report={offlineReadinessReport} />
-            </div>
-          ) : null}
-          <div className="crosshook-health-dashboard-expanded-meta">
-            <span>
-              <strong>Launch Method:</strong> {report.launch_method}
-            </span>
-            {meta?.last_success && (
-              <span>
-                <strong>Last Success:</strong> {formatRelativeTime(meta.last_success)}
-              </span>
-            )}
-            {meta != null && meta.failure_count_30d > 0 && (
-              <span>
-                <strong>Failures (30d):</strong> {meta.failure_count_30d}
-              </span>
-            )}
-            {meta?.is_favorite && <span>★ Favorite</span>}
-            {meta?.is_community_import && <span className="crosshook-status-chip">Community</span>}
-          </div>
-          {report.issues.length === 0 ? (
-            <p className="crosshook-muted">No issues found.</p>
-          ) : (
-            <ul className="crosshook-health-dashboard-issues-list">
-              {report.issues.map((issue) => {
-                const issueCategory = categorizeIssue(issue);
-                const protonField = issueCategory === 'missing_proton' ? getProtonPathField(issue.field) : null;
-                const isMigrationActive = activeMigrationField === issue.field;
-                const suggestion: MigrationSuggestion | null =
-                  protonField != null && scanResult != null
-                    ? (scanResult.suggestions.find((s) => s.profile_name === report.name && s.field === protonField) ??
-                      null)
-                    : null;
-                const hasNoMatch = protonField != null && scanResult != null && suggestion === null;
-
-                return (
-                  <li
-                    key={`${issue.field}-${issue.path}-${issue.message}-${issue.severity}`}
-                    className="crosshook-health-dashboard-issue"
-                  >
-                    <span className="crosshook-health-dashboard-issue__field">{issue.field}</span>
-                    {issue.path && <code className="crosshook-health-dashboard-issue__path">{issue.path}</code>}
-                    <span className="crosshook-health-dashboard-issue__message">{issue.message}</span>
-                    {issue.remediation && (
-                      <span className="crosshook-health-dashboard-issue__remediation crosshook-muted">
-                        {issue.remediation}
-                      </span>
-                    )}
-                    {issueCategory === 'missing_proton' && (
-                      <div className="crosshook-health-dashboard-migration-panel">
-                        {!isMigrationActive && (
-                          <button
-                            type="button"
-                            className="crosshook-button crosshook-button--small crosshook-focus-ring crosshook-nav-target"
-                            style={{ minHeight: 'var(--crosshook-touch-target-min)' }}
-                            onClick={() => void handleUpdateProton(issue.field)}
-                            disabled={isScanning}
-                            aria-label={`Update Proton for ${report.name}`}
-                          >
-                            Update Proton
-                          </button>
-                        )}
-                        {isMigrationActive && (
-                          <>
-                            {isScanning && <span className="crosshook-muted">Scanning Proton installations…</span>}
-                            {!isScanning && suggestion && (
-                              <div className="crosshook-health-dashboard-migration-inline">
-                                <div className="crosshook-health-dashboard-migration-path-comparison">
-                                  <code style={{ color: 'var(--crosshook-color-danger)' }}>{suggestion.old_path}</code>
-                                  {' \u2192 '}
-                                  <code style={{ color: 'var(--crosshook-color-success)' }}>{suggestion.new_path}</code>
-                                </div>
-                                <div className="crosshook-health-dashboard-migration-versions crosshook-muted">
-                                  {suggestion.old_proton_name} {'\u2192'} {suggestion.new_proton_name}
-                                </div>
-                                {suggestion.crosses_major_version && (
-                                  <div
-                                    role="alert"
-                                    className="crosshook-health-dashboard-migration-warning"
-                                    style={{ color: 'var(--crosshook-color-warning)' }}
-                                  >
-                                    &#9888; Major version change — WINE prefix may need recreation
-                                  </div>
-                                )}
-                                {!suggestion.crosses_major_version && suggestion.confidence < 0.75 && (
-                                  <div
-                                    role="alert"
-                                    className="crosshook-health-dashboard-migration-warning"
-                                    style={{ color: 'var(--crosshook-color-warning)' }}
-                                  >
-                                    &#9888; Different Proton family — verify compatibility
-                                  </div>
-                                )}
-                                <div className="crosshook-health-dashboard-migration-actions">
-                                  <button
-                                    type="button"
-                                    className="crosshook-button crosshook-button--small crosshook-focus-ring crosshook-nav-target crosshook-focus-target"
-                                    style={{ minHeight: 'var(--crosshook-touch-target-min)' }}
-                                    onClick={() => void handleApply(suggestion)}
-                                    disabled={isApplying}
-                                    aria-label={`Use ${suggestion.new_proton_name} for ${report.name}`}
-                                  >
-                                    {isApplying ? 'Updating\u2026' : `Use ${suggestion.new_proton_name}`}
-                                  </button>
-                                  <button
-                                    type="button"
-                                    className="crosshook-button crosshook-button--small crosshook-button--ghost crosshook-focus-ring crosshook-nav-target"
-                                    style={{ minHeight: 'var(--crosshook-touch-target-min)' }}
-                                    onClick={() => setActiveMigrationField(null)}
-                                    aria-label="Cancel Proton update"
-                                  >
-                                    Cancel
-                                  </button>
-                                </div>
-                              </div>
-                            )}
-                            {!isScanning && hasNoMatch && (
-                              <div className="crosshook-health-dashboard-migration-no-match">
-                                <span className="crosshook-muted">No matching Proton installation found.</span>{' '}
-                                <button
-                                  type="button"
-                                  className="crosshook-button crosshook-button--small crosshook-button--ghost crosshook-focus-ring crosshook-nav-target"
-                                  style={{ minHeight: 'var(--crosshook-touch-target-min)' }}
-                                  onClick={() => void onFixNavigate(report.name)}
-                                  aria-label={`Browse Proton path for ${report.name}`}
-                                >
-                                  Browse\u2026
-                                </button>
-                              </div>
-                            )}
-                            {migrationError && (
-                              <div
-                                role="alert"
-                                className="crosshook-health-dashboard-migration-error"
-                                style={{ color: 'var(--crosshook-color-danger)' }}
-                              >
-                                {migrationError}
-                              </div>
-                            )}
-                            {applyResult?.outcome === 'failed' && applyResult.profile_name === report.name && (
-                              <div
-                                role="alert"
-                                className="crosshook-health-dashboard-migration-error"
-                                style={{ color: 'var(--crosshook-color-danger)' }}
-                              >
-                                {applyResult.error ?? 'Failed to update Proton path.'}
-                              </div>
-                            )}
-                          </>
-                        )}
-                      </div>
-                    )}
-                  </li>
-                );
-              })}
-            </ul>
-          )}
-          {successNotice && (
-            <div
-              role="status"
-              className="crosshook-health-dashboard-migration-success"
-              style={{
-                color: 'var(--crosshook-color-success)',
-                display: 'flex',
-                alignItems: 'center',
-                gap: '8px',
-                marginTop: '8px',
-              }}
-            >
-              <span>&#10003; {successNotice}</span>
-              <button
-                type="button"
-                className="crosshook-button crosshook-button--small crosshook-button--ghost crosshook-focus-ring crosshook-nav-target"
-                style={{ minHeight: 'var(--crosshook-touch-target-min)' }}
-                onClick={() => setSuccessNotice(null)}
-                aria-label="Dismiss notification"
-              >
-                &#x2715;
-              </button>
-            </div>
-          )}
-          <div className="crosshook-health-dashboard-expanded-actions">
-            <button
-              type="button"
-              className="crosshook-button crosshook-button--small"
-              onClick={() => onRevalidate(report.name)}
-              aria-label={`Re-check ${report.name}`}
-            >
-              Re-check
-            </button>
-            {report.status !== 'healthy' && (
-              <button
-                type="button"
-                className="crosshook-button crosshook-button--small"
-                onClick={() => void onFixNavigate(report.name)}
-                aria-label={`Fix ${report.name}`}
-              >
-                Fix
-              </button>
-            )}
-          </div>
-        </div>
-      </td>
-    </tr>
-  );
-}
-
-function CommunityImportHealthPanel({ profiles }: { profiles: EnrichedProfileHealthReport[] }) {
-  const unhealthyImports = useMemo(
-    () => profiles.filter((r) => r.metadata?.is_community_import === true && r.status !== 'healthy'),
-    [profiles]
-  );
-
-  return (
-    <CollapsibleSection title="Community Import Health" defaultOpen={false}>
-      {unhealthyImports.length === 0 ? (
-        <p className="crosshook-muted">All community-imported profiles are healthy.</p>
-      ) : (
-        <>
-          <p className="crosshook-help-text crosshook-muted">
-            Imported profiles often need path adjustments for your system.
-          </p>
-          <ul className="crosshook-health-dashboard-panel-list">
-            {unhealthyImports.map((report) => (
-              <li key={report.name} className="crosshook-health-dashboard-panel-row">
-                <span className="crosshook-health-dashboard-panel-row__name">{report.name}</span>
-                <HealthBadge report={report} />
-                <span className="crosshook-muted">
-                  {report.issues.length} issue{report.issues.length !== 1 ? 's' : ''}
-                </span>
-              </li>
-            ))}
-          </ul>
-        </>
-      )}
-    </CollapsibleSection>
-  );
-}
+import { CommunityImportHealthPanel } from './health-dashboard/CommunityImportHealthPanel';
+import type { CardTrend, SortDirection, SortField, StatusFilter } from './health-dashboard/constants';
+import { STATUS_RANK, VERSION_STATUS_RANK } from './health-dashboard/constants';
+import { IssueBreakdownPanel, IssueDetailRow } from './health-dashboard/IssueBreakdownPanel';
+import { LauncherDriftPanel } from './health-dashboard/LauncherDriftPanel';
+import { RecentFailuresPanel } from './health-dashboard/RecentFailuresPanel';
+import { SkeletonCards, SummaryCard } from './health-dashboard/SummaryCards';
+import { SortArrow, TableToolbar } from './health-dashboard/TableControls';
+import {
+  categorizeIssue,
+  getVersionStatusColor,
+  getVersionStatusLabel,
+  mergeOfflineReadinessForRow,
+  offlineSortScore,
+} from './health-dashboard/utils';
 
 export function HealthDashboardPage({ onNavigate }: { onNavigate?: (route: AppRoute) => void }) {
   const { summary, loading, error, batchValidate, cachedSnapshots, revalidateSingle, trendByName, staleInfoByName } =
@@ -890,7 +96,7 @@ export function HealthDashboardPage({ onNavigate }: { onNavigate?: (route: AppRo
     }
   }
 
-  function handleRowClick(report: EnrichedProfileHealthReport) {
+  function handleRowClick(report: { name: string }) {
     setExpandedProfile((prev) => (prev === report.name ? null : report.name));
   }
